pyproject.toml
[tool.poetry]
name = "drf_test"
version = "0.1.0"
description = ""
authors = ["lolek"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
uritemplate = "^4.1.1"
PyYAML = "^6.0"
coreapi = "^2.3.3"
celery = "^5.2.7"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
